Authorized edit and delete question using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -84,7 +84,11 @@ class QuestionController extends Controller
     public function edit(Question $question)
     {
         //
+        if(\Gate::denies('update-question', $question)){ 
+            abort(403,"Access denied");
+        }
         return view("questions.edit", compact('question')); 
+        
     }
 
     /**
@@ -97,6 +101,9 @@ class QuestionController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+        if(\Gate::denies('update-question', $question)){ 
+            abort(403,"Access denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "The question has been updated");
     }
@@ -110,6 +117,9 @@ class QuestionController extends Controller
     public function destroy(Question $question)
     {
         //
+        if(\Gate::denies('update-question', $question)){ 
+            abort(403,"Access denied");
+        }
         $question->delete();
         return redirect('/questions')->with('success', "The question already deleted");
     }
